python/OcrScannerPdfToText/requirements.txt
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==12.0.0